refactors user data code, fixes folder name issue and removes some required fields on UserForm
