+++ frontend/src/components/UserForm/UserForm.js
@@ -8,11 +8,13 @@ const UserForm = () => {
   const formRef = useRef()
   const { updateUser } = useMutate()
   const { id } = useParams()
-  const user = useUser({ "id": id })
-  const isAdmin = user?.data?.data?.data.user.admin
+  const { status, data: userData } = useUser({ id })
 
-  const fetchedCountry = user?.data?.data?.data?.user.country;
-  const fetchedState = user?.data?.data?.data?.user.state;
+  const user = userData?.data?.data?.user
+  const isAdmin = user?.admin
+
+  const fetchedCountry = user?.country
+  const fetchedState = user?.state
   const [country, setCountry] = useState("");
   const [region, setRegion] = useState("");
 
@@ -21,8 +23,8 @@ const UserForm = () => {
     if (fetchedState) setRegion(fetchedState);
   }, [fetchedCountry, fetchedState]);
 
-  if (user.isError) return <h1>Something went wrong!</h1>
-  if (user.isLoading) return <h1>Loading...</h1>
+  if (status === "error") return <h1>Something went wrong!</h1>
+  if (status === "loading") return <h1>Loading...</h1>
 
   const handleSubmit = async (e) => {
     e.preventDefault()
@@ -58,17 +60,16 @@ const UserForm = () => {
           First Name:
           <input
             className="form-input admin-form-input"
-            defaultValue={user?.data?.data?.data?.user.firstName && user.data.data.data.user.firstName}
+            defaultValue={user.firstName}
             type="text"
             name="firstName"
-            required
           />
         </label>
         <label className="admin-form-label">
           Middle Name:
           <input
             className="admin-form-input"
-            defaultValue={user?.data?.data?.data?.user.middleName && user.data.data.data.user.middleName}
+            defaultValue={user.middleName}
             type="text"
             name="middleName"
           />
@@ -77,27 +78,25 @@ const UserForm = () => {
           Last Name:
           <textarea
             className="admin-form-input"
-            defaultValue={user?.data?.data?.data?.user.lastName && user.data.data.data.user.lastName}
+            defaultValue={user.lastName}
             type="text"
             name="lastName"
-            required
           />
         </label>
         <label className="admin-form-label">
           Gender:
           <textarea
             className="admin-form-input"
-            defaultValue={user?.data?.data?.data?.user.gender && user.data.data.data.user.gender}
+            defaultValue={user.gender}
             type="text"
             name="gender"
-            required
           />
         </label>
         <label className="admin-form-label">
           Email:
           <textarea
             className="admin-form-input"
-            defaultValue={user?.data?.data?.data?.user.email && user.data.data.data.user.email}
+            defaultValue={user.email}
             type="text"
             name="email"
             required
@@ -114,30 +113,27 @@ const UserForm = () => {
           Address1:
           <textarea
             className="admin-form-input"
-            defaultValue={user?.data?.data?.data?.user.address1 && user.data.data.data.user.address1}
+            defaultValue={user.address1}
             type="text"
             name="address1"
-            required
           />
         </label>
         <label className="admin-form-label">
           Address2:
           <textarea
             className="admin-form-input"
-            defaultValue={user?.data?.data?.data?.user.address2 && user.data.data.data.user.address2}
+            defaultValue={user.address2}
             type="text"
             name="address2"
-            required
           />
         </label>
         <label className="admin-form-label">
           City:
           <textarea
             className="admin-form-input"
-            defaultValue={user?.data?.data?.data?.user.city && user.data.data.data.user.city}
+            defaultValue={user.city}
             type="text"
             name="city"
-            required
           />
         </label>
         <label className="admin-form-label">
@@ -148,7 +144,6 @@ const UserForm = () => {
             country={country}
             value={region}
             onChange={(val) => setRegion(val)}
-            required
           />
         </label>
         <label className="admin-form-label">
@@ -158,27 +153,24 @@ const UserForm = () => {
             name="country"
             value={country}
             onChange={setCountry}
-            required
           />
         </label>
         <label className="admin-form-label">
           Zipcode:
           <textarea
             className="admin-form-input"
-            defaultValue={user?.data?.data?.data?.user.zipcode && user.data.data.data.user.zipcode}
+            defaultValue={user.zipcode}
             type="text"
             name="zipcode"
-            required
           />
         </label>
         <label className="admin-form-label">
           Phone Number:
           <textarea
             className="admin-form-input"
-            defaultValue={user?.data?.data?.data?.user.phoneNumber && user.data.data.data.user.phoneNumber}
+            defaultValue={user.phoneNumber}
             type="text"
             name="phoneNumber"
-            required
           />
         </label>
         <input className="admin-form-submit" type="submit" value="Submit" />
