+++ frontend/src/routes/AdminRoutes.js
@@ -6,7 +6,7 @@ import { useBoundStore } from "../stores/useBoundStore"
 import AdminNav from "../components/Admin/AdminNav/AdminNav"
 import AdminTitleBar from "../components/Admin/AdminTitleBar/AdminTitleBar"
 import UsersPage from "../pages/UsersPage/UsersPage"
-import UpdateUser from "../pages/CreateUpdateUser/UpdateUser"
+import UpdateUser from "../pages/UpdateUser/UpdateUser"
 import TagsPage from "../pages/TagsPage/TagsPage"
 import CreateUpdateTag from "../pages/CreateUpdateTag/CreateUpdateTag"
 import ProductsPage from "../pages/ProductsPage/ProductsPage"
